rename passes from cx to gate
