+++ qiskit/transpiler/passes/__init__.py
@@ -114,7 +114,7 @@ Additional Passes
 
    CheckMap
    CheckCXDirection
-   CXDirection
+   GateDirection
    MergeAdjacentBarriers
    BarrierBeforeFinalMeasurements
    RemoveFinalMeasurements
@@ -183,8 +183,10 @@ from .scheduling import TimeUnitAnalysis
 
 # additional utility passes
 from .utils import CheckMap
-from .utils import CheckCXDirection
-from .utils import CXDirection
+from .utils import CheckCXDirection  # Deprecated
+from .utils import CXDirection  # Deprecated
+from .utils import CheckGateDirection
+from .utils import GateDirection
 from .utils import BarrierBeforeFinalMeasurements
 from .utils import RemoveFinalMeasurements
 from .utils import MergeAdjacentBarriers
