+++ qiskit/transpiler/passes/utils/__init__.py
@@ -13,8 +13,10 @@
 """Utility passes used for other main passes."""
 
 from .check_map import CheckMap
-from .check_cx_direction import CheckCXDirection
-from .cx_direction import CXDirection
+from .check_cx_direction import CheckCXDirection  # Deprecated
+from .cx_direction import CXDirection  # Deprecated
+from .check_gate_direction import CheckGateDirection
+from .gate_direction import GateDirection
 from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
 from .remove_final_measurements import RemoveFinalMeasurements
 from .merge_adjacent_barriers import MergeAdjacentBarriers
